requirements.txt
about-time==4.2.1
alive-progress==3.3.0
attrs==25.4.0
autograd==1.8.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cma==4.4.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
Deprecated==1.3.1
filelock==3.20.1
fonttools==4.61.1
fsspec==2025.12.0
graphemeu==0.7.2
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
ml_dtypes==0.5.4
moocore==0.1.10
mpmath==1.3.0
natsort==8.4.0
networkx==3.4.2
ninja==1.13.0
nncf==2.19.0
numpy==2.2.6
onnx==1.20.0
openvino==2025.4.1
openvino-genai==2025.4.1.0
openvino-telemetry==2025.2.0
openvino-tokenizers==2025.4.1.0
optimum==2.1.0
optimum-intel==1.27.0
optimum-onnx==0.1.0
packaging==25.0
pandas==2.3.3
pillow==12.0.0
platformdirs==4.5.1
protobuf==6.33.2
psutil==7.2.1
pycparser==2.23
pydot==3.0.4
Pygments==2.19.2
pymoo==0.6.1.6
pyparsing==3.3.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rich==14.2.0
rpds-py==0.30.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.16.3
six==1.17.0
sympy==1.14.0
tabulate==0.9.0
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.9.1
tqdm==4.67.1
transformers==4.57.3
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
wrapt==2.0.1

# Stage 7: 推理設置所需的額外套件
accelerate==1.0.0           # 加速模型加載
python-dotenv==1.0.0        # 加載環境變數配置
hf_xet==0.1.8              # 加速 Hugging Face 下載（可選優化）

# Stage 8: 大型模型下載所需的額外套件
huggingface_hub==0.36.0     # 從 HuggingFace Hub 下載模型
